pidns: support external PID namespaces
This allows CRIU to restore a process into an existing PID namespace.
During checkpointing the PID namespace can be marked as external using:

   --external pid[<inode>]:<label>

This can be the host PID namespace or any other PID namespace.

During restore a process can be restored into an existing PID namespace
using:

  --inherit-fd fd[<FD>]:<label>

The <label> has to be the same for checkpoint and restore. CRIU uses
the <label> to know which resource the user means.

A process can start in the host PID namespace and can be moved to
another namespace or the other way around. Any PID namespace can be
used.

This is necessary to checkpoint containers in a POD which share certain
namespaces and this code to support external PID namespaces is the first
step towards checkpointing and restoring containers which belong to a
POD.

This is not using the --join-ns functionality as it is not possible to
move existing process to a PID namespace using setns(). Only child
process will be created in the PID namespace and that is why setns()
needs to be called before clone().

Signed-off-by: Adrian Reber <areber@redhat.com>